+++ style/index.js
@@ -1,1 +1,3 @@
+import 'reveal.js/reveal.css';
+import 'reveal.js/theme/white.css';
 import './base.css';